requirements.txt
aiohttp
pyyaml
attrs
openapi-spec-validator
strict_rfc3339
